backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
selenium==4.15.2
webdriver-manager==4.0.1
openai==1.6.1
beautifulsoup4==4.12.2
requests==2.31.0
selenium
requests
beautifulsoup4